+++ changelog.md
@@ -32,6 +32,7 @@
 - widget constructor can be deactivated in config page in order to save ressources
 - new version number format (e.g. v2.9.2) is standard in communication w/ io-drivers, old format is deprecated
 - menu button for German "Kurzanleitung" (separately available on github.com/smartvisu-newstuff/kurzanleitung) integrated in system menu
+- language is switched to "en" by default if configured language file does not exist any more
 
 ### Improvements
 - function Date().duration changed to accept negative values (for plots reaching into the future)